requirements.txt
azure-core==1.32.0
azure-identity==1.19.0
msal==1.31.1
msal-extensions==1.2.0
requests==2.32.3
Flask == 3.1.0